[crashlog][tests] Add basic integration test for kernel crashlog
Add an integration test that verifies basic behavior of stowing and
retrieving the kernel crashlog.  The test boots a qemu instance,
triggers a crash, which generates and stows a crashlog.  The system
then reboots and on the subsequent boot the crashlog is retrieved and
checked (only the most basic checks are peformed).

Updated the qemu integration test package to support arm64.

Changed "wait for qemu to start up" logic to wait for a newly enabled,
arch agnostic qemu trace event.  Previously, we waited for the
"SeaBIOS" message, which is x64 specific.

Currently, this test only works on arm64.  It's skipped on x64.  See
fxb/52843 for details.

A future CL will build on and extend this test infrastructure.

Bug: 52182
Test: out/core.arm64/host_x64/kernel_crashlog_tests
Change-Id: I28fc27af92661ec1b29bd0bf83f8f1fe411c8e5a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391509
Reviewed-by: George Kulakowski <kulakowski@google.com>
Reviewed-by: Scott Graham <scottmg@google.com>
Reviewed-by: Francois Rousseau <frousseau@google.com>
Commit-Queue: Nick Maniscalco <maniscalco@google.com>
Testability-Review: Scott Graham <scottmg@google.com>
Testability-Review: Francois Rousseau <frousseau@google.com>