[kernel][x86] idle_thread: Don't send an IPI if a target is spinning
When a target CPU is spinning before going to sleep, a waking CPU does
not need to send an IPI to wake the target; it can use an interlocked
CMPXCHG to mark the target as woken and avoid the IPI. This reduces the
number of unnecessary IPI (APIC interactions), hardware interrupts,
and halt instructions; those operations are particularly expensive in
virtual machines, so this slightly speeds up certain Fuchsia tests in
VMs.

This optimization only applies to the non-MWAIT path; the MWAIT path
has a very similar interlock and we can combine them in a future CR,
as we move spin and halt into a set of choices for a unified sleep
predictor.

Tested:
1. Booted on a GCP nested VM. Reduced test run time of fidl_compatibility
   test by a further 10%.

Bug: 49133 profile fyi builders timing out with ...
Bug: 33981 Fuchsia should run on Intel NUC6CAYH
Bug: 37450 Fuchsia should run well on AMD Ryzen (17h), ZP-B1 and ZP/PiR-B2
Bug: 33638 Fuchsia should run (well) on grunt (AMD) chromebook
Change-Id: I63ca5e52eae7169a5d03bb7c818c293090c78f9a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391677
Reviewed-by: Corey Tabaka <eieio@google.com>
Testability-Review: Corey Tabaka <eieio@google.com>
Commit-Queue: Venkatesh Srinivas <venkateshs@google.com>