[virtcon] Fix keep-log-visible for run-vc terminals
The implementation of the fuchsia.virtualconsole.SessionManager service
ignored the kernel virtcon.keep-log-visible flag. This meant that if the
virtual consoles were instantiated by the appmgr that flag would have no
effect.

With this change a call to SessionManager.CreateSession will switch to
the new console session unless the virtcon.keep-log-visible is set and
the currently active console is the log.

Bug: 52926
Test: booted with and without virtcon.keep-log-visible=true

Change-Id: I14d4703ad54d82abf7bda8db920f2668694c8097
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393282
Testability-Review: Ashley Coleman <ashleycoleman@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Carl Norum <cjn@google.com>