[camera][controller] Ensure that driver callbacks are posted as tasks on the loop
Prior to this CL, the callbacks from drivers like GDC/GE2D and ISP
executed partially on the drivers thread and partially on the
controller's loop.

This CL ensures that the callbacks are posted as tasks on the
controller's loop.

Test: Added asserts which fails without this change while running
config cycling on camera-gym for a long time.

Change-Id: Ia0e1c539e4333981b619515ff05f33d703cbbfa9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393156
Reviewed-by: Adam MacBeth <amacbeth@google.com>
Reviewed-by: Matt Sandy <msandy@google.com>
Commit-Queue: Brijen Raval <braval@google.com>
Testability-Review: Matt Sandy <msandy@google.com>