[a11y] A11yManager waits to be fully-initialized before requestingg settings.
This change prepares the A11yManager to initialize the Screen Reader with its
correct locale. Because it is not possible to init the Screen Reader without
this information, A11YManager waits to receive this info before requesting
settings.

Requesting settings when the A11yManager is not initialized may lead in a race
condition between receiving the locale and receiving the settings.

Change-Id: I2f5fc587d3a80492dcf8caa3632435b0060eb2ba
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393500
Reviewed-by: Alice Neels <neelsa@google.com>
Testability-Review: Alice Neels <neelsa@google.com>
Commit-Queue: Lucas Radaelli <lucasradaelli@google.com>