[pager] Supply uncommitted pages
Extends the semantics of zx_pager_supply_pages to allow the aux_vmo to
not have to be fully committed. Pages that are not committed are
notionally zero anyway, and therefore have a well defined value in the
target pager_vmo. This can be done efficientally without allocating
pages in the target through the use of markers.

Allowing uncommitted pages in the aux_vmo provides flexibility in how
the kernel chooses to represent VMOs, and specifically allows it to
merge committed zero pages back to the global zero page. Without this
change there is no way to reliably provide a page of zeros by supply
pages, as you are racing with the kernel zero page scanner.

Bug: 52652

Change-Id: I805e39559aec3c5b588bd78483520f72b0cdcd7c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392980
Testability-Review: Adrian Danis <adanis@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: Rasha Eqbal <rashaeqbal@google.com>