[kernel][timer] Make TimerQueue::TransitionOffCpu more symmetric
Originally, this transition code was a function which took two
cpu_num_t parameters to identify the source and the destination of the
transition.

This patch restores the parity between the target (now, the this
pointer on the member function) and the source (the parameter) by
identifying the later directly, rather than via a cpu_num_t.

Change-Id: Iefeafb64462a8f81c636fe5b681e37c7f87a3a5f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394459
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Testability-Review: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: George Kulakowski <kulakowski@google.com>