[ffx] slightly improve internal error output
All errors that are not handled by higher level subcommands, that
reach main will be considered a bug. We format the error output a
litlte more, and include the causation trail, while also telling the
user that this is a bug.

Change-Id: I9e66e251fe9cbd2f0edea44a55b0b83be1ec8fa0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393818
Reviewed-by: Jordon Wing <jwing@google.com>
Commit-Queue: James Tucker <raggi@google.com>
Testability-Review: James Tucker <raggi@google.com>