[cmc] Replace cml::Ref with contextual types
This change adds a procedural macro to generate reference types, since
there are a lot of them. It replaces the `Ref` type with these types. We
introduce an `AnyRef` type which is similar to `Ref`, but it is not
deserializable: it is only for use in helper functions that want to
operate on multiple reference types.

This change is a first step towards replacing cml_schema.json with serde
validation. Instead of validating the references with a regex, we can
rely on the serde deserializer to report an error. This change doesn't
enable that validation yet, though.

Bug: 52963

Change-Id: If2d90c8734023b1f994132be52a8a4251895fb0f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392302
Reviewed-by: Adam Perry <adamperry@google.com>
Reviewed-by: Adam Lesinski <adamlesinski@google.com>
Testability-Review: Adam Perry <adamperry@google.com>
Testability-Review: Adam Lesinski <adamlesinski@google.com>
Commit-Queue: Gary Bressler <geb@google.com>