[oom][logs] Drop "OOM" prefix from debug logs
All the logs we print from the oom-thread have the "OOM" prefix, which
can be misleading. The presence of the word "OOM" seems to indicate that
the we're running out of memory, which can be confusing if the log line
printed out was "OOM: memory availability state Normal".

Replace the "OOM" prefix with a more neutral "memory-pressure". Also
rename the "oom-thread" to "memory-pressure-thread".

Test: oom_tests

Change-Id: I57002d23743348be486200ff442256c2f3df503a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394793
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: Scott Graham <scottmg@google.com>
Testability-Review: Scott Graham <scottmg@google.com>