[audio_core] Pass VolumeCurve to OutputPipeline
The OutputPipeline will use this to convert gain to volume so that
effects can be informed of the current gain/volume. This is a
foundational change that handles the plumbing of this structure to make
the following functional changes clearer.

Bug: 43979
Test: No behavior change;
      fx test audio_core_unittests
Change-Id: I74e03eebf07ca4efed49c70387f02081b50a478b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394292
Reviewed-by: Tom Bergan <tombergan@google.com>
Testability-Review: Tim Detwiler <tjdetwiler@google.com>
Commit-Queue: Tim Detwiler <tjdetwiler@google.com>