[fidl][llcpp] remove Initialize indirection
This initialization is not needed because we just allocated the message.

This doesn't change the codesize because the compiler was smart enough
to see two initializations and to just keep one.

This cl will allow to remove some templates later.

Change-Id: Ib10b7343936a3e79cf96a29812b1b39257e3a78d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394739
Commit-Queue: Pascal Perez <pascallouis@google.com>
Reviewed-by: Yifei Teng <yifeit@google.com>
Reviewed-by: Pascal Perez <pascallouis@google.com>
Testability-Review: Yifei Teng <yifeit@google.com>