[brightness_manager] Fix bug with blue/green swap
The blue and green values were swapped on the service side
of the fidl interface to conform to HID spec, but broke the
parsing on this side. The blue green values are swapped now
in the parsing logic to correct the mismatch.

Test: fx test -o bright_manager_tests
Bug: b/157594412
Change-Id: Ia4ac2bd8f67758a4629497a4a45e2bd13453eea9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394693
Reviewed-by: Kevin Twidle <kpt@google.com>
Testability-Review: Chip Fukuhara <cfukuhara@google.com>
Commit-Queue: Paul Faria <paulfaria@google.com>