[flatland] Convert matrices to rectangles.
The conversion logic from matrices to rectangles was moved out of the
Renderer in change 391533. This change formalizes some of that logic
into a function to be called in the main render loop.

The intermdiate matrix format is not currently needed anywhere, but
will likely be necessary in the future for transforming input, so
this change leaves compute matrices and computing rectangles as
separate functions.

Rectangles are still computed and used for LinkSystem::UpdateLinks(),
which ensures that the Flatland API calls result in matrices that are
valid for rectangle conversion.

Change-Id: I9c1e2b35d8ecc91421f15d8bfe14a068e712acc1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393067
Commit-Queue: Evan Birenbaum <birenbaum@google.com>
Testability-Review: Evan Birenbaum <birenbaum@google.com>
Reviewed-by: Adam Gousetis <adamgousetis@google.com>