[build][testing] Update OWNERship of test_spec+
It makes less sense of me to own this these days - and increasingly more
sense for the individuals I've replaced my name with in this change.
Moreover, likely better for this to be maintained by those who work
closer to and think more about the platform-infra divide.

Change-Id: I0e3a36083987f30b966cd1bf4142bd78fcce1141
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395614
Reviewed-by: Shai Barack <shayba@google.com>
Reviewed-by: Gary Miguel <garymm@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>