[pci] Extend BusLinkInterface into BusDeviceInterface
Devices need a way to request the bus driver allocate
their MSI objects, so extend BusLinkInterface to be
BusDeviceInterface and now reflect an interface for
devices to request resources and device linkage from
the parent bus. This is also required for GetBti which
is addressed in a later CL. Pulling BDI out of the bus
header was necessary to allow a proper dependency chain
with the fake test objects.

bug: 32978
Change-Id: I26893b5214b7df0303a5c70cfb7ba4e29a2d3059
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393734
Commit-Queue: Christopher Anderson <cja@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Testability-Review: Adrian Danis <adanis@google.com>