[carnelian] Refactor Scenic support
Refactor scenic support as a general cleanup and to more easily add
better frame scheduling later.

The goal was to move all scenic specific code into scenic modules. This
CL doesn't quite get there, do to the need to forward scenic input
events to the app strategy for processing, but it is a long way towards
it. Refactoring the input handling feels like too big a step for this
CL.

Moved the app strategies into their own modules to make the separation
easier to maintain.

Got rid of the supports_scenic method on AppStrategy as it is no longer
needed. Instead there is a method for creating a view at launch time
for testing. This is only implemented on Scenic, as Framebuffer always
creates a view at launch.

Initialization order has changed enough that I was able to remove the
unnecessary use of Option for any fields in App. That structure is now
fully initialized on creation.

ViewController now has a single constructor function which uses the
view strategy for behavior differences on Scenic and Framebuffer.

Got rid of the ScenicResources struct, as it no longer needs to be used
from multiple places. Now its fields are part of ScenicViewStrategy.

Removed the unnecessary _async name suffix on functions that used to
have a sync and async version.

Testing: all automated tests, plus manual testing of all examples,
recovery and terminal.

Change-Id: I5c0273a83933c47c961c0d168fc7b2deaa058da5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392808
Reviewed-by: David Reveman <reveman@google.com>
Testability-Review: David Reveman <reveman@google.com>
Commit-Queue: Rob Tsuk <robtsuk@google.com>