[sysmem-make-tracking] sysmem::MakeTracking<>()
This is mostly for setting of table fields without
needing to redundantly specify the type of a value
that's already the same type as the field being
assigned/moved to the field.

We're considering doing something similar for
fidl::Allocator::make<T>() so that T can be inferred,
but we're removing that from the critical path for
now.

If setting a table field which is itself a table,
the table Frame will be allocated, which is useful
for incremental building.

Change-Id: Ie59a254019f7b3c367e4f1da65d533b3cdc49b69
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/388124
Commit-Queue: Dustin Green <dustingreen@google.com>
Reviewed-by: Benjamin Prosnitz <bprosnitz@google.com>
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Testability-Review: Benjamin Prosnitz <bprosnitz@google.com>