[sme] [brcmfmac] Handle channel change indication for SoftAP.
Channel change indication is issued when the firmware changes
the SoftAP's channel (for example, the client interface in the same
device associates to an AP running on a different channel).

Test: Added a unit test to check the handling of this event.
Added another test to verify softap creation failure in the driver,
but this test needs the sim timer changes.
Also, verified on astro and sherlock that the channel indication is
handled. The test sequence is:
- Start SoftAP on channel x
- Check wlan status reports SoftAP is running on channel x
- Associate client IF to an AP on channel y
- Check wlan status reports SoftAP is running on channel y

Change-Id: I1ee90cdb8666b348766bda8d676a4d5ec920420d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/383932
Commit-Queue: Karthik Krishnaswami <karthikrish@google.com>
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Reviewed-by: Eric Wang <eyw@google.com>
Testability-Review: Zhiyi Chen <zhiyichen@google.com>