[covargs] Remap source files to new checkout.
The coverage recipe creates a new checkout to run with the covargs tool
which has a different path than the checkout that was actually used to
build. We need to remap the file paths to use the new checkout so it
can properly find the source files and upload their corresponding html
files with the coverage report.

Test: https://chromium-swarm.appspot.com/task?id=4c95a9df8a55d410

Bug: 49473
Change-Id: I31dc789a52e9319ce5e8160cf587bf2511ab11cd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394179
Testability-Review: Ina Huh <ihuh@google.com>
Commit-Queue: Ina Huh <ihuh@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>