[bt][a2dp-source] Select correct bitpool for SBC
SBC negotiation would select 53 as a constant bitpoolvalue whenever we
selected SBC, which happened occasionally.  It is not required that
sinks support this value, and sometimes have a smaller maximum value.

Select the maximum value that is supported if it is below 53.  Also
change the default value to 51, which is recommended for Joint Stereo
48kHz audio by the A2DP spec (Sec 4.3.2.6).

Fixed: 52376

Test: updated test_stream_selection in bt-a2dp-source-tests

Multiply: bt-a2dp-source-tests
Change-Id: Ida2d0bf00d768d417d7c7d40969da419fe630ad7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393729
Commit-Queue: Marie Janssen 💖 <jamuraa@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>
Testability-Review: Austin Foxley <afoxley@google.com>