[ffx] Begin internal documentation.
This is a good start.  There still needs to be some documentation around
the persistent configuration files, how to setup an environment, etc.
Also, how to use the config macros in the library.  But since this was
timeboxed, this is what I have so far.

Change-Id: I920f31571ae272a8e5087f555929f7152103e439
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394187
Commit-Queue: Matthew Boetger <boetger@google.com>
Reviewed-by: Jordon Wing <jwing@google.com>