[ffx] Detect DaemonProxy param in ffx_plugin
With the double-dispatch it's now possible to identify that a ffx_plugin
takes a DaemonProxy instead of a RemoteControlProxy.

Extending this method in future CLs should allow for ffx_plugins that
take a number of arbitrary Proxies, and then use the new RCS selector
interface to gather the proxies - generating all the boilerplate code.

Change-Id: If52821aed9697706b151a70c86cdc8facdbe6b58
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/393195
Commit-Queue: Matthew Boetger <boetger@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>
Testability-Review: Andrew Davies <awdavies@google.com>