[settings] Add InspectBroker to write all setting values to inspect
The InspectBroker sits between the registry and setting handlers. It listens for the Restore command on boot and asks each setting handler for its value when it replies to the Restore command. When a setting handler tells the registry its value changed, the InspectBroker also requests the value. This isolates all the inspect-related logic in a single broker and no other pieces of the settings service need to worry about it.

Bug: 51123
Fixes: 44874
Multiply: setui_service_tests
Test: fx test -o setui_service_tests
Change-Id: Id81a7937c6cd694417f84069f3bb31be7f330bff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/384690
Commit-Queue: William Xiao <wxyz@google.com>
Testability-Review: William Xiao <wxyz@google.com>
Reviewed-by: Paul Faria <paulfaria@google.com>
Reviewed-by: Chip Fukuhara <cfukuhara@google.com>