[devices] Do not explicitly install libdriver in bootfs.
Now that no bootfs content comes from the ZN build anymore, libdriver always
naturally gets pulled into zbi's via the inclusion of a driver which depends
on it.

Adding libdriver directly is problematic as doing so completely bypasses
variant selection which is done at the executable/driver level. The current
zbi building step does not rely on the dependency tree in order to identify
what should be included in the zbi, but its replacement will. The dependency
removed in this change would have had the consequence of always including
the no-variant version of libdriver (and some shared libraries it depends on)
in zbi's, which is inaccurate and wasteful.

Bug: 45680
Change-Id: I4b4744e544847cb9a6800bbc5e7e3fb5808c9d65
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395773
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: P.Y. Laligand <pylaligand@google.com>