[fbl][lists] Add split_after to lists.
Add a split_after operation to the single and doubly linked list
containers.  The operation accepts either an iterator or direct
reference to an element in the list (call it A), and splits the list
in two lists at the point after A.  It will return a new list whose
head is the element immediately following A, while A will become the
tail of the list which was split.

The split_after operation is O(1).  Sized lists do not support split
after as maintenance of the size member would require O(n)
complexity.

Tests: new unit tests
Fixed: 53051
Change-Id: Ibe89ae2fad153a4083b5bb2136faecbc79c279ce
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396033
Commit-Queue: John Grossman <johngro@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>
Testability-Review: George Kulakowski <kulakowski@google.com>