[blobfs][pager] Pass const reference to UserPagerInfo
The user pager functions need to consult a blob's |UserPagerInfo| for
various bits of information required on the paging path. Instead of
passing in a |UserPagerInfo*|, pass in a |const UserPagerInfo&|. Once
the |UserPagerInfo| is created at the time of initialization of the
blob, it is never modified. Besides, these functions should not be
tampering with the contents of the |UserPagerInfo| anyway.

Test: No behavior change. Existing blobfs unit & integration tests pass.

Change-Id: Iee49979042bd5334c4cfb4c514e663f1d0eb09f1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396055
Commit-Queue: Rasha Eqbal <rashaeqbal@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>
Testability-Review: James Sullivan <jfsulliv@google.com>