[fshost] Fix block watcher race.
Opening the block device before taking the lock and checking if the
watcher was paused could sometimes result in a race where the watcher
would be resumed after the block device was added, but before we'd
opened it, causing the tests to flake. Fix this by acquiring the lock
before opening the block device.

Bug: 53368
Change-Id: I72ae883be9e17c27b55502417d6a91f5f8164394
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395758
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Testability-Review: Suraj Malhotra <surajmalhotra@google.com>
Commit-Queue: Simon Shields <simonshields@google.com>