[fshost] Bump pkgfs startup deadline
We are investigating a possible performance regression in pkgfs in
which it takes ~30 seconds to start on emulators when running an SDK
companion image. Bump the timeout to 40 seconds while we investigate
both the performance of pkgfs startup and removing this timeout
entirely.

Bug: 53639
Change-Id: I3bd9ff02d92d33161cbd8348622c49e378021c9c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396849
Reviewed-by: James Sullivan <jfsulliv@google.com>
Reviewed-by: Manali Bhutiyani <manalib@google.com>
Testability-Review: James Sullivan <jfsulliv@google.com>
Commit-Queue: John Wittrock <wittrock@google.com>