[build][fuzzing] Fix fdio build for fuzzer variants
libfdio.so was recently migrated from ZN to GN as part of
fxrev.dev/392329. Unfortunately, that CL broke the workaround for
fxbug.dev/38226. This wasn't caught by CQ as fuzzer builds have been
previously included in CQ. (See fxbug.dev/53411 for the request to
change that.)

This CL includes the workaround for fxb/38226 into
//build/images/finalize_manifest.py, so that the generated package
manifests correctly indicate to use the uninstrumented libfdio.so.

It also updates the "--fuzz-with" option to `fx set` to enable the
variant toolchain needed to build the sanitized-but-not-covered
libfdio.so.

Bug: 38226, 53028, 53411
Change-Id: I360a381e39f4a46dfc059b21841c3927067c0870
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394198
Reviewed-by: P.Y. Laligand <pylaligand@google.com>
Reviewed-by: David Turner <digit@google.com>
Reviewed-by: Marco Vanotti <mvanotti@google.com>
Commit-Queue: Aaron Green <aarongreen@google.com>
Testability-Review: Aaron Green <aarongreen@google.com>