[appmgr][cmx] Propagate all entries in "program" to runner
This allows component manifests to specify additional attributes in the
"program" field for the runner to use in executing the component. This
is closer to the component framework V2 definition, which passes all
fields in the corresponding dictionary to the runner to interpret. The
attributes are distinct from the "args" value, which are for the use
of the component being run.  Only string values are permitted to
reflect fuchsia.sys.ProgramMetadata definition.

This updates appmgr to propagate these fields through and updates the
cmx_schema to allow additional properties in the "program" object so
long as the values are all strings.

Bug: 53243

Change-Id: Ie9b86ee213fcb7105a16be709641af27c33debf9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396269
Commit-Queue: James Robinson <jamesr@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Testability-Review: Gary Bressler <geb@google.com>