[audio][astro] Change PDM paramters for 96kHz
We change the PDM CIC filtering parameters so instead of halfing the
last downsample rate in the chain, we half the first one and adjust the
gain accordingly as seen in a reference implementation. This improves
the performance seen in capturing 28kHz tones in Astro, and shows no
adverse effect on lower frequencies.

We also update a HPF parameter difference from the reference
implementation.

This change is now virtually identical to Eric's
https://fuchsia-review.googlesource.com/c/fuchsia/+/386033

Bug: 53207
Test: Manual recording with audio-driver-ctl with lower (16kHz tone and regular music) and higher (28kHz tone) frequencies.

Change-Id: Icd93a0bebdab22ca56174b4b018a3feba8d94328
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395676
Reviewed-by: Eric Holland <hollande@google.com>
Reviewed-by: Tom Bergan <tombergan@google.com>
Reviewed-by: Tim Detwiler <tjdetwiler@google.com>
Testability-Review: Andres Oportus <andresoportus@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>