[debugger] DieRef->SymbolRef, make Kind an enum.
Renames IndexNode::DieRef to IndexNode::SymbolRef, and changes the "is
declaration" flag in that structure to an enum.

This change was started to allow adding ELF symbols to the index. The
more pressing issue requiring that was solved a different way so I will
not be following up with more patches along those lines soon. But we
will likely want to add ELF symbols to the index in the future which
makes this generalization still desirable.

There should be no behavior change.

Bug: 53091

Change-Id: Ia07205d3875dc79a4761d3a484d461fded6438bf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392533
Reviewed-by: Dangyi Liu <dangyi@google.com>
Testability-Review: Brett Wilson <brettw@google.com>
Commit-Queue: Brett Wilson <brettw@google.com>