[scenic] Fix ViewTree error message for root view_tree
There's an error message in ViewTree that fires when a Session has more
than one connected view. Because the ViewTree also counts the Scene node
as equivalent to a view this message also fires if the root session
contains a view. This CL fixes that, allowing the root session to own
a view. Necessary for input injection.

Test: No functional changes. Manually tested.
Change-Id: Ibeaed7c170f62009dcb7f5d76733399f324b19b4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397034
Reviewed-by: Evan Birenbaum <birenbaum@google.com>
Testability-Review: Evan Birenbaum <birenbaum@google.com>
Commit-Queue: Mikael Pessa <mikaelpessa@google.com>