[zx] Fix zx::vmar argument order to match C interface
When we updated the C interface for zx::vmar to match the API rubric, we
did not update the C++ interface. This CL begins the process of updating
the C++ interface to match the C interface by introducing new methods
with the proper argument order.

For zx::vmar::map, we can add the new method with the final name because
the zx::vmo argument will disamiguate the overloads. For
zx::vmar::protect and zx::vmar::allocate, an actual parameter of 0 would
be ambiguous, so we need to go through a more complicated sequence of
introducing a protect2 and allocate2 with the proper argument order,
removing the old methods, and then migrating back to the original names
with the proper argument order.

Bug: 52432
Change-Id: I9b89d69c996ff13db40b03c4df7942037958853d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/394178
Commit-Queue: Adam Barth <abarth@google.com>
API-Review: Adam Barth <abarth@google.com>
Testability-Review: Adam Barth <abarth@google.com>
Reviewed-by: George Kulakowski <kulakowski@google.com>