[wlan][factory] Combine and increase buffer size
The factory tool uses the same buffer as input and output so it would be
eaiser if the size of request and result vectors match.

Also increase the size to 16k because it is the largest command buffer
defined anywhere in the factory tool.

Test: None. There is no functional change.
Change-Id: I95b148c2eb827c20d3d46332352d2a2b72156c4d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396077
Reviewed-by: Sakthi Vignesh Radhakrishnan <rsakthi@google.com>
Reviewed-by: Karthik Krishnaswami <karthikrish@google.com>
API-Review: Rebecca Silberstein <silberst@google.com>
Testability-Review: Eric Wang <eyw@google.com>
Commit-Queue: Eric Wang <eyw@google.com>