[Fuchsia DevTools] add --once option to Harvester
The Harvester is normally a long running process. For some testing purposes it
is helpful to run the update loop one time and exit. This CL adds a command line
option for `--once` that will run the main update loop one time and then exit.

Bonus: adds a join on the secondary thread so that it shuts down without
segfaulting when the main thread exits. This bug was exposed by adding the
--once option.

Test: src/tests/end_to_end/harvester_test.dart
Change-Id: Ifd4a7a65a0e31dbb8e6b850d8a2f98580578a4fe
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/391756
Reviewed-by: Isma Cortes <isma@google.com>
Reviewed-by: Yuan Zhi <yuanzhi@google.com>
Reviewed-by: Jeremy Manson <jeremymanson@google.com>
Reviewed-by: Jayson Adams <shrike@google.com>
Commit-Queue: Dave Schuyler <dschuyler@google.com>
Testability-Review: Jeremy Manson <jeremymanson@google.com>