[audio][sherlock] Allow and warn ring buffer position too big
Assertion for ring buffer position too big from the HW added
in https://fuchsia-review.googlesource.com/c/fuchsia/+/378538
shows that it is possible to get a report from the HW that is
beyond the ring buffer size. This CL changes the assertion to
a warning and a reset of the position to 0.

Fixed: 52004
Test: audio_driver_tests
Multiply: audio_driver_tests
Change-Id: If77248af9221970707da0c03cec81434ad7f4347
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/392774
Reviewed-by: Eric Holland <hollande@google.com>
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>
Testability-Review: Braden Kell <bradenkell@google.com>