[affectedtests] Reland affectedtests speedup
Now with 100% less flakes.
This was previously flaky because the logic to shard the
Ninja graph in its dotfile representation wasn't stable.
As a result we would fail to acknowledge some of the nodes
or edges, and then miss some possible paths in the graph.

Multiply: affectedtests_tests

Change-Id: I16702dabce1aab2c929edc1ccc6dec0beff670b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/397044
Testability-Review: Shai Barack <shayba@google.com>
Commit-Queue: Shai Barack <shayba@google.com>
Reviewed-by: Gary Miguel <garymm@google.com>