[scrutiny] Integrate Plugin Management, Data Model and REST service.
This change integrates three core concepts to the Scrutiny auditing
framework:

1. Plugins: The ability to register and load plugins dynamically.
2. Data Model: The core data intake and control interfaces for the system.
3. REST Service: The JSON REST HTTP interface to the data controllers which
   visualization tools will interface with.

This CL defines each of the core layers of the framework and integrates
a simple way to add new "plugins" to the system. Plugins are designed to
be loaded and unloaded allowing for different configurations to be
loaded based on the auditing task at hand.

This allows plugin authors to develop complex, resource intensive
plugins and load them on demand without having to worry as much about
overall system impact. This should enable much more specialized tools to
be developed and shared without having to worry as much about their
generality as they can be disabled by default.

At its core a plugin is just a way to manage a set of DataCollectors and
DataControllers. This provides a familiar interface to new plugin
authors allowing them to quickly get something up and running with
minimal boilerplate. Registering the controllers and running the
collectors it 100% handled by the PluginManager engine allowing
developers to spend more time working directly with the abstract data
model.

The REST interface is automatically generated based on the plugins
DataControllers namespaces and so no boilerplate is required to be
written by plugin developers further streamlining development.

Change-Id: I90e2603cb00e6bb10964a39f71c4bf6156777bdf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/389350
Reviewed-by: Bryan Henry <bryanhenry@google.com>
Reviewed-by: Benjamin Wright <benwright@google.com>
Reviewed-by: Daniel Hwang <arkay@google.com>
Testability-Review: Bryan Henry <bryanhenry@google.com>
Commit-Queue: Benjamin Wright <benwright@google.com>