[ulib][zxc] Add nodiscard attribute to zx::status and fitx::result.
The C++17 "[[nodiscard]]" attribute gives a warning if a function returns
a zx::status or fitx::result, but the caller ignores the result.
Ignoring the result is often indicative of a bug, where the caller
hasn't carried out error handling.

In cases where the error should be ignored, casting the result to void
will suppress the warning, as follows:

    zx::status<> MountFilesystem();

    void Init() {
      MountFilesystem();  // warning emitted

      zx::status<> status = MountFilesystem();  // no warning
      if (status.is_error()) {
         // ...
      }

      (void)MountFilesystem();  // no warning
    }

When the library is being compiled with C++14, we fall back to
__attribute__((__warn_unused_result__)) if possible, or otherwise just
ignore the attribute.

Change-Id: If2a37facaf68fd6cfdd94f585ab1dbe4a88f9ded
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395763
Commit-Queue: David Greenaway <dgreenaway@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Testability-Review: David Greenaway <dgreenaway@google.com>