[cm][test_utils] Adds BlackBoxTestBuilder
With the need to add an additional argument to the launched
component manager for BlackBoxTest, some callsites would
need to be updated to pass an empty vector.

A better pattern is to use a Builder to configure the BlackBoxTest
without needing to specify all parameters.

Change-Id: I9aae944044cda54c737e682b5edd8d6d49ff3f5d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/396107
Commit-Queue: Adam Lesinski <adamlesinski@google.com>
Reviewed-by: Fady Samuel <fsamuel@google.com>
Testability-Review: Fady Samuel <fsamuel@google.com>