[mdns] remove check for config dir existence
A check was recently added to make sure the config directory exists
before calling JSONParser::ParseFromDirectory. As it turns out, that
method was supposed to handle non-existent directories without
indicating an error, so the correct fix was to fix the method. That
change is made by fxr/395654. This CL reverses the mdns change,
leaving the relevant test in place.

Test: fx test mdns_tests
      Unit test for the non-existent directory case remains.
Change-Id: I7fbca4558136dff9ed60902dd16bd83dc2b4ce5f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/395673
Testability-Review: Dale Sather <dalesat@google.com>
Reviewed-by: Austin Foxley <afoxley@google.com>